requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
argparse=1.4.0=pypi_0
attrs=23.2.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bs4=0.0.2=pypi_0
bzip2=1.0.8=h6c40b1e_5
ca-certificates=2024.3.11=hecd8cb5_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
dataclasses-json=0.6.4=pypi_0
deprecated=1.2.14=pypi_0
dirtyjson=1.0.8=pypi_0
distro=1.9.0=pypi_0
expat=2.5.0=hcec6c5f_0
filelock=3.13.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
greenlet=3.0.3=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.4=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.21.4=pypi_0
idna=3.6=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.4=hecd8cb5_0
llama-index=0.10.20=pypi_0
llama-index-agent-openai=0.1.6=pypi_0
llama-index-cli=0.1.10=pypi_0
llama-index-core=0.10.20.post2=pypi_0
llama-index-embeddings-huggingface=0.1.4=pypi_0
llama-index-embeddings-openai=0.1.6=pypi_0
llama-index-indices-managed-llama-cloud=0.1.4=pypi_0
llama-index-legacy=0.9.48=pypi_0
llama-index-llms-openai=0.1.12=pypi_0
llama-index-multi-modal-llms-openai=0.1.4=pypi_0
llama-index-program-openai=0.1.4=pypi_0
llama-index-question-gen-openai=0.1.3=pypi_0
llama-index-readers-file=0.1.11=pypi_0
llama-index-readers-llama-parse=0.1.3=pypi_0
llama-index-vector-stores-pinecone=0.1.4=pypi_0
llama-parse=0.3.9=pypi_0
llamaindex-py-client=0.1.13=pypi_0
markupsafe=2.1.5=pypi_0
marshmallow=3.21.1=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=pypi_0
nltk=3.8.1=pypi_0
numpy=1.26.4=pypi_0
openai=1.14.2=pypi_0
openssl=3.0.13=hca72f7f_0
packaging=24.0=pypi_0
pandas=2.2.1=pypi_0
pillow=10.2.0=pypi_0
pinecone-client=3.1.0=pypi_0
pip=23.3.1=py312hecd8cb5_0
pydantic=2.6.4=pypi_0
pydantic-core=2.16.3=pypi_0
pymupdf=1.23.26=pypi_0
pymupdfb=1.23.22=pypi_0
pypdf=4.1.0=pypi_0
python=3.12.2=hd58486a_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=hca72f7f_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
safetensors=0.4.2=pypi_0
setuptools=68.2.2=py312hecd8cb5_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.28=pypi_0
sqlite=3.41.2=h6c40b1e_0
striprtf=0.0.26=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
tiktoken=0.6.0=pypi_0
tk=8.6.12=h5d9f67b_0
tokenizers=0.15.2=pypi_0
torch=2.2.1=pypi_0
tqdm=4.66.2=pypi_0
transformers=4.38.2=pypi_0
typing-extensions=4.10.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
wheel=0.41.2=py312hecd8cb5_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h6c40b1e_0
yarl=1.9.4=pypi_0
zlib=1.2.13=h4dc903c_0
